rename detector_plotter file to detectorplotter for consistency
