+++ detectorplotter.py
