+++ example_instruments/sans2d/SANS2D_example.py
@@ -1,6 +1,6 @@
 from collections import OrderedDict
 from nexusbuilder import NexusBuilder
-from detector_plotter import DetectorPlotter
+from detectorplotter import DetectorPlotter
 
 if __name__ == '__main__':
     output_filename = 'SANS_example_gzip_compress.hdf5'
